+++ packages/core/src/serializer.ts
@@ -20,14 +20,46 @@ export interface PrimitiveTypeSerializer {
     serializeValue(value: unknown, property: Property): string | undefined
 }
 
+const isPrimitiveTypeSerializer = (value: unknown): value is PrimitiveTypeSerializer =>
+       typeof value === "object"
+    && value !== null
+    && "serializeValue" in value
+    && typeof value.serializeValue === "function"
+
+
+/**
+ * Type to provide options to the serializer.
+ */
+export type SerializationOptions = Partial<{
+    /**
+     * Determines whether empty feature values are skipped during serialization.
+     * Default = false, meaning that empty feature values are *not* skipped.
+     * (The specification states that empty feature values SHOULD be serialized,
+     * but not that they MUST be.)
+     */
+    skipEmptyValues: boolean
+
+    primitiveTypeSerializer: PrimitiveTypeSerializer
+}>
+
+
 /**
  * @return a {@link SerializationChunk} of the given model (i.e., an array of {@link Node nodes} - the first argument) to the LionWeb serialization JSON format.
  */
 export const serializeNodes = <NT extends Node>(
     nodes: NT[],
     extractionFacade: ExtractionFacade<NT>,
-    primitiveTypeSerializer: PrimitiveTypeSerializer = new DefaultPrimitiveTypeSerializer()
+    primitiveTypeSerializerOrOptions?: PrimitiveTypeSerializer | SerializationOptions
 ): SerializationChunk /* <=> JSON */ => {
+    const primitiveTypeSerializer = (
+        isPrimitiveTypeSerializer(primitiveTypeSerializerOrOptions)
+            ? primitiveTypeSerializerOrOptions
+            : primitiveTypeSerializerOrOptions?.primitiveTypeSerializer
+    ) ?? new DefaultPrimitiveTypeSerializer()
+    const skipEmptyValues = isPrimitiveTypeSerializer(primitiveTypeSerializerOrOptions)
+        ? false
+        : primitiveTypeSerializerOrOptions?.skipEmptyValues
+
     const serializedNodes: SerializedNode[] = []  // keep nodes as much as possible "in order"
     const ids: { [id: string]: boolean } = {}   // maintain a map to keep track of IDs of nodes that have been serialized
     const languagesUsed: Language[] = []
@@ -65,7 +97,11 @@ export const serializeNodes = <NT extends Node>(
                 version: featureLanguage.version,
                 key: feature.key
             }
-            if (feature instanceof Property && value !== undefined) {
+            if (feature instanceof Property) {
+                if (value === undefined) {
+                    // for immediate backward compatibility: skip empty property values regardless of options?.skipEmptyValues
+                    return
+                }
                 const encodedValue = (() => {
                     // (could also just inspect type of value:)
                     if (feature.type instanceof PrimitiveType) {
@@ -87,6 +123,9 @@ export const serializeNodes = <NT extends Node>(
             }
             if (feature instanceof Containment) {
                 const children = asArray(value) as (NT | null)[]
+                if (children.length === 0 && skipEmptyValues) {
+                    return
+                }
                 serializedNode.containments.push({
                     containment: featureMetaPointer,
                     children: asIds(children)
@@ -103,6 +142,9 @@ export const serializeNodes = <NT extends Node>(
             if (feature instanceof Reference) {
                 // Note: value can be null === typeof unresolved, e.g. on an unset (or previously unresolved) single-valued reference
                 const targets = asArray(value) as (NT | null)[]
+                if (targets.length === 0 && skipEmptyValues) {
+                    return
+                }
                 serializedNode.references.push({
                     reference: featureMetaPointer,
                     targets: targets